Break up output package
